requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scipy>=1.11

# ML
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
btyd>=0.1

# LLM
groq>=0.4
openai>=1.0
huggingface-hub>=0.20
transformers>=4.36
sentence-transformers>=2.2

# NLP
nltk>=3.8
rapidfuzz>=3.5

# Visualization
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18

# Web App
streamlit>=1.30

# MLOps
mlflow>=2.10

# Database
sqlalchemy>=2.0

# Caching
diskcache>=5.6

# Testing
pytest>=7.4
pytest-cov>=4.1

# Code Quality
flake8>=6.1
black>=23.12

# Utilities
pyyaml>=6.0
python-dotenv>=1.0
faker>=22.0
